mcp-tasks/pyproject.toml
[project]
name = "fastmcp-task-manager"
version = "0.1.0"
description = "Task manager MCP (FastMCP + SQLite)"
requires-python = ">=3.11"
dependencies = [
  "fastmcp>=2.0.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "pydantic-ai>=1.79.0",
  "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
  "e2e: curl-пробы к поднятому mcp-tasks по HTTP (нужен MCP_TASKS_E2E_URL)",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
